Change minZoom for subway
hey there Brandon et al - I'm a huge fan of the project.

I'd like to propose reducing the `minZoom`  for the subway road-layer data. My motivation is to see a city's subway network from the zoom level of a normal city.

Please let me know if this change is appropriate, or can be improved in any way. I know there are a lot of plans for future transit data. This data is not currently included in the default basemap.
cheers